+++ src/test/config.test.ts
@@ -1,14 +1,17 @@
 import {expect} from 'chai';
 import {dirname} from 'path';
 
-import {getConfig} from '../config';
+
 import {displayAllLoggers} from '../logging';
+import {getConfig} from './util';
 
 
-const configFilepath = process.cwd() + '/fixtures/.vcms.yml';
+const defaultConfigFilepath = __dirname + '/../../fixtures/.vcms.yml';
+const defaultStartupConfigScriptFilepath = __dirname;
+
 
 suite('Config', () => {
-  const debug = () => {
+  const log = () => {
     displayAllLoggers();
   };
 
@@ -16,37 +19,22 @@ suite('Config', () => {
     displayAllLoggers(false);
   });
 
-  const run = async (args: any, configFilepath?: string) => {
-    // save originals
-    const originalArgv = process.argv;
-
-    // change the execution context
-    process.argv = ['node', 'app'].concat(args);
-
-    // run update and get config
-    const config = await getConfig(configFilepath);
-
-    // get back to the original context
-    process.argv = originalArgv;
-
-    // return the result
-    return config;
-  };
 
   test('defaults', async () => {
-    const config = await run([]);
+    const config = await getConfig();
     expect(config.PORT).to.equal(8000);
+    expect(config.publicDirectory).to.equal('public');
   });
 
   test('change port', async () => {
-    const config = await run(['-p', '123']);
+    const config = await getConfig(['-p', '123']);
     expect(config.PORT).to.equal(123);
   });
 
   test('config file takes precedence over defaults', async () => {
     // default port is 8000, the port in the file is 8080 for dev
-    const config = await run([], configFilepath);
-    expect(config.PORT).to.equal(8080);
+    const config = await getConfig([], null, defaultConfigFilepath);
+    expect(config.PORT).to.equal(123);
   });
 
 
@@ -57,7 +45,7 @@ suite('Config', () => {
     const originalEnvPort = process.env.PORT;  // save context
     process.env.PORT = '321';                  // change context
 
-    const config = await run([], configFilepath);
+    const config = await getConfig([], null, defaultConfigFilepath);
     expect(config.PORT).to.equal(321);
 
     if (!originalEnvPort) delete process.env.PORT;  // restore context
@@ -71,7 +59,7 @@ suite('Config', () => {
     const originalEnvPort = process.env.PORT;  // save context
     process.env.PORT = '321';                  // change context
 
-    const config = await run(['-p', '4444'], configFilepath);
+    const config = await getConfig(['-p', '4444'], null, defaultConfigFilepath);
     expect(config.PORT).to.equal(4444);
 
 
@@ -84,43 +72,30 @@ suite('Config', () => {
 
   test('process.env.NODE_ENV influences the configuration', async () => {
     const originalNODE_ENV = process.env.NODE_ENV;  // save context
-    process.env.NODE_ENV = 'dev';                   // change context
 
+    process.env.NODE_ENV = 'dev';  // change context
     // dev port is 3001
-    let config = await run([], configFilepath);
+    let config = await getConfig([], null, defaultConfigFilepath);
     expect(config.PORT).to.equal(3001);
 
 
     process.env.NODE_ENV = 'prod';  // change context
     // prod port is 8080
-    config = (await run([], configFilepath));
+    config = await getConfig([], null, defaultConfigFilepath);
     expect(config.PORT).to.equal(8080);
 
-    process.env.NODE_ENV = originalNODE_ENV;  // restore context
-  });
-
-
-
-  let TITLE =
-      'process.env.NODE_ENV influenced configuration takes precedence over node-env influenced configuration';
-  test(TITLE, async () => {
-    // make sure NODE_ENV influences the configuration
-    const originalNodeEnv = process.env.NODE_ENV;  // save context
-    process.env.NODE_ENV = 'prod';                 // change context
-
-
-    // the default file has a "dev" node-env with port 3001
-    // but NODE_ENV is "prod" and "prod" node-env has 8080 port
-    let config = await run([], configFilepath);
-    expect(config.PORT).to.equal(8080);
-
-    // restore context
-    if (originalNodeEnv) process.env.NODE_ENV = originalNodeEnv;
+    // restore defaults
+    if (!originalNODE_ENV) {
+      delete process.env.NODE_ENV;
+    } else {
+      process.env.NODE_ENV = originalNODE_ENV;
+    }
   });
 
 
   test('changing the public directory', async () => {
-    let config = await run(['--public-directory', 'hello'], configFilepath);
+    let config = await getConfig(
+        ['--public-directory', 'hello'], null, defaultConfigFilepath);
     expect(config.publicDirectory).to.equal('hello');
   });
 });