+++ test/app.js
@@ -1,22 +1,4 @@
 const {start} = require('../lib/vcms');
 
 
-const config = {
-  configFilepath: `${__dirname}/.vcms.yml`,
-  routers: {
-    '/example': require('./example.router'),
-    '/session': require('./session.router')
-  },
-  initSessionFunction: async (session) => {
-    if (!session.hello) {
-      session.hello = 'hello world from session'
-    }
-  },
-  middlewares: [(req, res, next) => {
-    console.log(`asking for ${req.url} (log from user-defined middleware)`);
-    next();
-  }]
-}
-
-// start the server
-start(config);
+start();