+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # ChangeLog
 
+## [16-05-2018] v1.1.0
+
+* adding startup configuration script support.
+* adding `util` test module in npm libraries.
+* update readme
+
 ## [15-05-2018] v1.0.6
 
 * logging errors if any happens at start time.